Update XML Model to simplify the renderer a bit.
This adds a specialization of XmlElement called XmlElementWithChildren
that holds the child elements. XmlElement no longer can hold child
elements.

This demonstrates the design concept of using specialized objects to
remove "if" statements. It is usually true that we know how an object
will be used when it is constructed. Many "if" statements are used to
discover the state of an object. If we used specialized objects, we know
what their state will be.